app/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
fastapi_mail
python-multipart
starlette 
# amqp==5.0.6
xlrd
Babel
pyjwt
# Celery==5.1.2 
jinja2
Pillow
twilio
numpy 
pandas 
boto3 
passlib 
aiofiles 
fastapi 
uvicorn 
gunicorn 
jmespath 
websockets 
botocore 
s3transfer 
SQLAlchemy 
apscheduler 
fastapi-mail 
cryptography 
python-dotenv 
psycopg2-binary 
python-dateutil 
python-multipart 
SQLAlchemy-Utils 
typing_extensions 
redis 
rq 
uuid 
xattr 
http-ece 
py-vapid 
pywebpush 
broadcaster 
asyncio-redis